cascading v1.0

+++ GA/individual.h
@@ -8,9 +8,9 @@ using namespace std;
 class Individual
 {
 public:
-	unsigned long int genotype[DIM];
-	double phenotype[DIM];
-	double funcVal;
+	int genotype[DIM];
+	int phenotype[DIM];
+	int funcVal;
 	double fitness;
 	bool isparent;
 	bool is_chosen;
@@ -20,7 +20,7 @@ public:
 public:
 	Individual(Functions &func);
 	Individual(Functions &func, double phenotype[DIM]);
-	Individual(Functions &func, unsigned long int genotype[DIM]);
+	Individual(Functions &func, int genotype[DIM]);
 	Individual(Functions &func, Individual* individual);
 	void Fit();
 	void Encode();
